feat: adicionado limitador para visibilidade do problema para 5 casos de teste validados

+++ app/Observers/SubmitRunObserver.php
@@ -27,7 +27,8 @@ class SubmitRunObserver
      */
     public function deleted(SubmitRun $submitRun): void
     {
-        $submitRun->file->delete();
+        if(!!$submitRun->file_id)
+            $submitRun->file->delete();
     }
 
     /**
